for second review

+++ src/scripts/index.js
@@ -1,5 +1,5 @@
 import '../pages/index.css';
-import { createCard, likeCard } from './card';
+import { createCard, likeCard, getCardForDeletion } from './card';
 import { showModal, closeModal } from './modal';
 import { enableValidation, clearValidation } from './validation';
 import { getUserInfo, getCards, deleteMyCard, editProfile, addNewCard, changeAvatar } from './api';
@@ -19,15 +19,15 @@ const cardTemplate = document.querySelector('#card-template').content;
 const cardList = document.querySelector('.places__list');
 
 // Функция вставки/добавления карточки
-function addCard(item, cardList, cardTemplate, createCard, likeCard, openCard, deleteThisCard, profileId) {
-  const cardElement = createCard(item, cardTemplate, likeCard, openCard, deleteThisCard, profileId);
+function addCard(item, cardList, cardTemplate, createCard, likeCard, openCard, openDeletePopup, profileId) {
+  const cardElement = createCard(item, cardTemplate, likeCard, openCard, openDeletePopup, profileId);
   cardList.append(cardElement);
 };
 
 // Функция заполнения карточками страницы
 function fillCards(initialCards, profileId) {
   initialCards.forEach((card) => {
-    addCard(card, cardList, cardTemplate,createCard, likeCard, openCard, deleteThisCard, profileId);
+    addCard(card, cardList, cardTemplate, createCard, likeCard, openCard, openDeletePopup, profileId);
   });
 };
 
@@ -111,7 +111,7 @@ function handleAddForm(evt) {
   addNewCard(cardValue, linkValue)
     .then((card) => {
       console.log('Добавление карточки прошло успешно:', card.link);
-      const newCard = createCard(card, cardTemplate, likeCard, openCard, deleteThisCard, profileId);
+      const newCard = createCard(card, cardTemplate, likeCard, openCard, openDeletePopup, profileId);
       cardList.prepend(newCard);
       closeModal(addPopup);
     })
@@ -188,18 +188,42 @@ function handleProfileForm(evt) {
 
 profileForm.addEventListener('submit', handleProfileForm);
 
+// Модальное окно удаление карточки с сайта (DOM-элементы и переменные - открытие/закрытие попапа)
+const deletePopup = document.querySelector('.popup_type_delete');
+const closeDeleteButton = deletePopup.querySelector('.popup__close');
+const deleteForm = document.querySelector('form[name="delete-card"');
+
+const openDeletePopup = () => {
+  showModal(deletePopup);
+};
+
+const closeDeletePopup = () => {
+  closeModal(deletePopup);
+};
+
+closeDeleteButton.addEventListener('click', closeDeletePopup);
+
 // Функция удаления карточки
-function deleteThisCard(cardId, deleteButton) {
+function deleteThisCard({ cardId, deleteButton }) {
   deleteMyCard(cardId)
     .then((res) => {
       console.log('Удаление карточки произошло успешно:', res);
       const deleteItem = deleteButton.closest('.places__item');
       deleteItem.remove();
+      closeDeletePopup();
     })
     .catch((error) => {
       console.log('Ошибка при удалении карточки:', error);
     });
 };
+
+// Вызов функции удаления карточки при нажатии на кнопку - "Да"
+function handleDeleteForm(evt) {
+  evt.preventDefault();
+  deleteThisCard(getCardForDeletion());
+};
+
+deleteForm.addEventListener('submit', handleDeleteForm);
 
 // Вызов функции, отвечающей за включение валидации всех форм
 enableValidation(validationConfig);
